pyproject.toml
[project]
name = "rebaser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Ashlynn Antrobus", email = "ashrobertsantrobus@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "gitpython>=3.1.44",
    "google-genai>=1.8.0",
    "loguru>=0.7.3",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.0",
]

[project.scripts]
rebaser = "rebaser:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
